# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Forafix API Documentation",
        "description": "Marketplace API with Wallet, Escrow and Booking management",
        "contact": {
            "email": "support@forafix.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/email/verify/{id}/{hash}": {
            "get": {
                "tags": [
                    "Verification"
                ],
                "summary": "Verify email address via signed link",
                "operationId": "6ed540fa14ef7f273aa29e8927d30c69",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "hash",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to login"
                    }
                }
            }
        },
        "/api/email/resend": {
            "post": {
                "tags": [
                    "Verification"
                ],
                "summary": "Resend email verification link",
                "operationId": "43852a3070150f6782a13d075b970da4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Link sent"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}